Pull out bearer token class
