+++ LinkTwrapper.Domain/packages.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="System.Net.Http" version="2.0.20126.16343" targetFramework="net46" />
+  <package id="System.Net.Http.Formatting.Extension" version="5.2.3.0" targetFramework="net46" />
+</packages>
